get line

+++ libft/ft_substr.c
@@ -22,7 +22,7 @@ static char	*empty_string(void)
 	return (sub);
 }
 
-char	*ft_substr(char const *s, unsigned int start, unsigned int len)
+char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*sub;
 	size_t	i;
